Clean up code

+++ chia-protocol/src/fullblock.rs
@@ -43,7 +43,8 @@ impl FullBlock {
         self.foliage.foliage_transaction_block_hash.is_some()
     }
 
-    pub fn total_iters(&self) -> u64 {
+    #[cfg_attr(target_arch = "wasm32", wasm_patch::conv_u128_to_u64_for_wasm)]
+    pub fn total_iters(&self) -> u128 {
         self.reward_chain_block.total_iters
     }
 
@@ -51,7 +52,8 @@ impl FullBlock {
         self.reward_chain_block.height
     }
 
-    pub fn weight(&self) -> u64 {
+    #[cfg_attr(target_arch = "wasm32", wasm_patch::conv_u128_to_u64_for_wasm)]
+    pub fn weight(&self) -> u128 {
         self.reward_chain_block.weight
     }
 
@@ -112,7 +114,7 @@ impl FullBlock {
 
     #[getter]
     #[pyo3(name = "total_iters")]
-    fn py_total_iters(&self) -> u64 {
+    fn py_total_iters(&self) -> u128 {
         self.total_iters()
     }
 
@@ -124,7 +126,7 @@ impl FullBlock {
 
     #[getter]
     #[pyo3(name = "weight")]
-    fn py_weight(&self) -> u64 {
+    fn py_weight(&self) -> u128 {
         self.weight()
     }
